app/service/train_mleng/requirements.txt
google-cloud-storage
pandas==0.25.2
scikit-learn==0.21.3
xgboost==0.90
PyYAML==5.1